v2.10, v2.11
v2.10: Another delay option - slow down third step of chains when not using helix. Delays set to 4,5.5 and 7,8 atm, not tested. Good for LiqFus on HaughtyTulittia
v2.11: Removed helix block on debuff-sharing NMs if the force_allow_helix flag is set. No command for this one, set it manually in the xml file